# src/docs/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Kafka REST API SwaggerUI",
    "version": "0.1.0",
    "title": "Kafka REST API SwaggerUI",
    "contact": {
      "name": "gnuhpc",
      "url": "https://github.com/gnuhpc",
      "email": "gnuhpc@gmail.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
    "/jmx/v1": {
      "get": {
        "tags": [
          "collector-controller"
        ],
        "summary": "Fetch all JMX metric data",
        "operationId": "collectJMXMetricUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "jmxurl",
            "in": "query",
            "description": "Parameter jmxurl should be a comma-separated list of {IP:Port} or set to 'default'",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/JMXMetricDataV1"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/jmx/v2": {
      "post": {
        "tags": [
          "collector-controller"
        ],
        "summary": "Fetch JMX metric data with query filter. You can get the query filter template through the API /jmx/v2/filters.",
        "operationId": "collectJMXMetricUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "jmxurl",
            "in": "query",
            "description": "Parameter jmxurl should be a comma-separated list of {IP:Port} or set to 'default'",
            "required": false,
            "type": "string"
          },
          {
            "in": "body",
            "name": "jmxQuery",
            "description": "jmxQuery",
            "required": true,
            "schema": {
              "$ref": "#/definitions/JMXQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/JMXMetricData"
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/jmx/v2/filters": {
      "get": {
        "tags": [
          "collector-controller"
        ],
        "summary": "List the query filter templates with the filterKey. If filterKey is set to empty, it will return all the templates.",
        "operationId": "listJMXFilterTemplateUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "filterKey",
            "in": "query",
            "description": "filterKey",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "HashMap«string,object»": {
      "type": "object",
      "additionalProperties": {
        "type": "object"
      }
    },
...
    "JMXConfiguration": {
      "type": "object",
      "properties": {
        "include": {
          "$ref": "#/definitions/JMXFilter"
        },
        "exclude": {
          "$ref": "#/definitions/JMXFilter"
        }
      }
    },
    "JMXFilter": {
      "type": "object",
      "properties": {
        "attribute": {
          "type": "object"
        },
        "beanNames": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "beanRegexes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Pattern"
          }
        },
        "domain": {
          "type": "string"
        },
        "domainRegex": {
          "$ref": "#/definitions/Pattern"
        },
        "emptyBeanName": {
          "type": "boolean"
        },
        "filter": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        }
      }
    },
    "JMXMetricData": {
      "type": "object",
      "properties": {
        "collected": {
          "type": "boolean"
        },
        "host": {
          "type": "string"
        },
        "metrics": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/HashMap«string,object»"
          }
        },
        "msg": {
          "type": "string"
        },
        "timestamp": {
          "type": "string",
          "example": "yyyy-MM-dd HH:mm:ss"
        }
      }
    },
    "JMXMetricDataV1": {
      "type": "object",
      "properties": {
        "collected": {
          "type": "boolean"
        },
        "host": {
          "type": "string"
        },
        "mbeanInfo": {
          "type": "object"
        },
        "msg": {
          "type": "string"
        },
        "timestamp": {
          "type": "string",
          "example": "yyyy-MM-dd HH:mm:ss"
        }
      }
    },
    "JMXQuery": {
      "type": "object",
      "properties": {
        "filters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/JMXConfiguration"
          }
        }
      }
    },
...
    "Pattern": {
      "type": "object",
      "properties": {
        "cursor": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
  }
}